requirements-dev.txt
fastapi==0.104.1
pydantic-settings==2.1.0
python-multipart==0.0.6
uvicorn==0.24.0.post1
python-lsp-server[all]
pytest==7.2.0
pytest-cov==2.11.1
pytest-mock==3.10.0
debugpy==1.6.7
